Add files via upload
My Name is Anthony and this script is used to encrypt a message without importing any modules and functions in Python. It is a simple script that uses input, string slicing, lists, for loops , if statements, Dictionary and print functions. After learning about security, I wanted to try and build my own script to encrypt and decrypt and I believe this works as tested. I am open to feedback and suggestions.