Add files via upload
This script is used to encrypt a plain text file(.txt) in Python. It uses python commands such as input, string, import, slicing, lists, for loops , if statements, Dictionary and print functions. After learning about security, I wanted to try and build my own script to encrypt and decrypt files and this works as tested. I am open to feedback and suggestions.